Move all file gflags into gjslint.py (which will get replaced eventually by newstart.py).
